Converted request to typescript.
